chore: move some other things around a little bit
