+++ plugins/scaffolder-backend/package.json
@@ -64,9 +64,6 @@
     "@backstage/plugin-scaffolder-common": "workspace:^",
     "@backstage/plugin-scaffolder-node": "workspace:^",
     "@backstage/types": "workspace:^",
-    "@gitbeaker/core": "^35.6.0",
-    "@gitbeaker/node": "^35.1.0",
-    "@octokit/webhooks": "^10.0.0",
     "@types/express": "^4.17.6",
     "@types/luxon": "^3.0.0",
     "azure-devops-node-api": "^11.0.1",
@@ -83,14 +80,11 @@
     "isomorphic-git": "^1.23.0",
     "jsonschema": "^1.2.6",
     "knex": "^3.0.0",
-    "libsodium-wrappers": "^0.7.11",
     "lodash": "^4.17.21",
     "luxon": "^3.0.0",
     "morgan": "^1.10.0",
     "node-fetch": "^2.6.7",
     "nunjucks": "^3.2.3",
-    "octokit": "^2.0.0",
-    "octokit-plugin-create-pull-request": "^3.10.0",
     "p-limit": "^3.1.0",
     "p-queue": "^6.6.2",
     "prom-client": "^14.0.1",
@@ -101,12 +95,10 @@
     "zod": "^3.21.4"
   },
   "devDependencies": {
-    "@backstage/backend-test-utils": "workspace:^",
     "@backstage/cli": "workspace:^",
     "@types/command-exists": "^1.2.0",
     "@types/fs-extra": "^9.0.1",
     "@types/git-url-parse": "^9.0.0",
-    "@types/libsodium-wrappers": "^0.7.10",
     "@types/nunjucks": "^3.1.4",
     "@types/supertest": "^2.0.8",
     "@types/zen-observable": "^0.8.0",
@@ -115,8 +107,7 @@
     "msw": "^1.0.0",
     "strip-ansi": "^7.1.0",
     "supertest": "^6.1.3",
-    "wait-for-expect": "^3.0.2",
-    "yaml": "^2.0.0"
+    "wait-for-expect": "^3.0.2"
   },
   "files": [
     "dist",
