+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/publish/util.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,116 +14,7 @@
  * limitations under the License.
  */
 
-import { InputError } from '@backstage/errors';
-import { isChildPath } from '@backstage/backend-common';
-import { join as joinPath, normalize as normalizePath } from 'path';
-import { ScmIntegrationRegistry } from '@backstage/integration';
-
-export const getRepoSourceDirectory = (
-  workspacePath: string,
-  sourcePath: string | undefined,
-) => {
-  if (sourcePath) {
-    const safeSuffix = normalizePath(sourcePath).replace(
-      /^(\.\.(\/|\\|$))+/,
-      '',
-    );
-    const path = joinPath(workspacePath, safeSuffix);
-    if (!isChildPath(workspacePath, path)) {
-      throw new Error('Invalid source path');
-    }
-    return path;
-  }
-  return workspacePath;
-};
-export type RepoSpec = {
-  repo: string;
-  host: string;
-  owner?: string;
-  organization?: string;
-  workspace?: string;
-  project?: string;
-};
-
-export const parseRepoUrl = (
-  repoUrl: string,
-  integrations: ScmIntegrationRegistry,
-): RepoSpec => {
-  let parsed;
-  try {
-    parsed = new URL(`https://${repoUrl}`);
-  } catch (error) {
-    throw new InputError(
-      `Invalid repo URL passed to publisher, got ${repoUrl}, ${error}`,
-    );
-  }
-  const host = parsed.host;
-  const owner = parsed.searchParams.get('owner') ?? undefined;
-  const organization = parsed.searchParams.get('organization') ?? undefined;
-  const workspace = parsed.searchParams.get('workspace') ?? undefined;
-  const project = parsed.searchParams.get('project') ?? undefined;
-
-  const type = integrations.byHost(host)?.type;
-
-  if (!type) {
-    throw new InputError(
-      `No matching integration configuration for host ${host}, please check your integrations config`,
-    );
-  }
-
-  const repo: string = parsed.searchParams.get('repo')!;
-  switch (type) {
-    case 'bitbucket': {
-      if (host === 'www.bitbucket.org') {
-        checkRequiredParams(parsed, 'workspace');
-      }
-      checkRequiredParams(parsed, 'project', 'repo');
-      break;
-    }
-    case 'gitlab': {
-      // project is the projectID, and if defined, owner and repo won't be needed.
-      if (!project) {
-        checkRequiredParams(parsed, 'owner', 'repo');
-      }
-      break;
-    }
-    case 'gerrit': {
-      checkRequiredParams(parsed, 'repo');
-      break;
-    }
-    default: {
-      checkRequiredParams(parsed, 'repo', 'owner');
-      break;
-    }
-  }
-
-  return { host, owner, repo, organization, workspace, project };
-};
-export const isExecutable = (fileMode: number) => {
-  const executeBitMask = 0o000111;
-  const res = fileMode & executeBitMask;
-  return res > 0;
-};
-
-/**
- * This function checks if the required parameters (given as the `params` argument) are present in the URL
- *
- * @param repoUrl - the URL for the repository
- * @param params - a variadic list of URL query parameter names
- *
- * @throws
- * An InputError exception is thrown if any of the required parameters is not in the URL.
- *
- * @public
- */
-function checkRequiredParams(repoUrl: URL, ...params: string[]) {
-  for (let i = 0; i < params.length; i++) {
-    if (!repoUrl.searchParams.get(params[i])) {
-      throw new InputError(
-        `Invalid repo URL passed to publisher: ${repoUrl.toString()}, missing ${
-          params[i]
-        }`,
-      );
-    }
-  }
-}
+//   const executeBitMask = 0o000111;
+//   const res = fileMode & executeBitMask;
+//   return res > 0;
+// };
