+++ plugins/scaffolder-node/src/files/deserializeDirectoryContents.test.ts
