+++ plugins/scaffolder-node/src/files/deserializeDirectoryContents.ts
