+++ plugins/scaffolder-node/src/files/index.ts
