+++ plugins/scaffolder-node/src/files/serializeDirectoryContents.test.ts
