+++ plugins/scaffolder-node/src/files/serializeDirectoryContents.ts
